[B360M-55] Removing OpalImagePicker form Podfile + Updating file upload implementation
